pyproject.toml
[project]
name = "pyrizrenderer"
version = "0.1.0"
description = "添加你的描述这里"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "keyboard>=0.13.5",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pydub>=0.25.1",
    "pygame-ce>=2.5.5",
    "pyopengl>=3.1.9",
    "pywin32>=311",
    "tqdm>=4.67.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
